backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
pyjwt
joblib
numpy<2.0
scikit-learn==1.6.1
xgboost
shap
python-dotenv
pyjwt
google-generativeai
